Including maxWireVersion etc in ServerDescription encourages race.
The Server Discovery And Monitoring Spec showed the ServerDescription struct
with minWireVersion and maxWireVersion, and also encourages clients to store
maxDocumentSize, maxMessageSize, and maxWriteBatchSize there. But if clients
use these fields they're prone to races.

Guidance for determining these values belongs in no particular spec at the
moment, unfortunately. SDAM is a reasonable place for the warning to live.