Revert "Prefer local mongos"
Reverts 5f563d232d7be5596dc9ec5038bf22b1d7367915.

We've decided not to do this after all. We'll write a very broad, optional spec
to guide drivers that want to provide a "server selection callback" that
applications can register to customize which servers they prefer.

That'll allow people to implement the "prefer localhost" behavior if they want
- we can write good guides for them that include the information about using
getaddrinfo to recognize IPv6 addresses for localhost - as well as customers
with more complex needs.