fix(change-streams): change streams should close on missing token
If a user choses to use the change stream helper, and not just run
an aggregation (resulting in a raw cursor), then it is assumed that
they have opted in to "resumability".  If we receive a response doc
that does not have a resume token, we can no longer provide this
guarantee and therefore MUST close the stream.