test(max-staleness): make sure we are testing the correct value
The lack of a `maxWireProtocol` field in the ismaster on this test
forces an immediate incompatibility error in creation of a read
preference due to lack of a server greater than protocol 5. This
change ensures we are directly testing the invalid value of `1`
for the `maxStalesnessSeconds` value, rather than the aforementioned
incompatibility.