SPEC-1161 Test that txn reads don't inherit readConcern
The Transactions Spec says, "Drivers MUST override all other collection,
database, or client readConcerns with the transaction’s readConcern.
Drivers MUST add this readConcern to the first command in a transaction
if and only if the readConcern is supplied and not the default."

I believe this applies when a transaction has *no* readConcern as well:
drivers ignore collection, database, or client readConcern and use no
readConcern in this case.