DRIVERS-556: Convenient API for Transactions spec tests (#435)
* Add additional withTransaction tests

* Rename spec directory

* Test that commit retry only overwrites WC's w option

* Document clientOptions in test README

* withTransaction can still succeed if callback aborts

* Use consistent result.error fields from transactions spec tests

* Add wtimeout error test and fix other WC error fail points

* Reorganize callback/commit tests and add WC and transient error commit tests

* Add prose tests for callback error and return value behavior

* Transient error on commit should retry entire transaction

* Fix readConcern APM assertions

Per the txn spec, only the initial command in a transaction can include a readConcern. Furthermore, afterClusterTime would only be specified if the client session is re-used for a second transaction (only one of our tests).

* Use more robust assertions for extra ops after a transaction

Asserting a value for txnNumber assumes that retryWrites=true, which may not be the case. Instead, ignore that field entirely and just assert that autocommit is not found. Combined with asserting that startTransaction is not found, we effectively assert that the command is not part of a transaction.

* Expect abortTransaction after non-transient error

Per the spec, withTransaction should abort if the callback reports an error and the session is in the "starting" or "in progress" state.

* Skip commit-retry tests pending SPEC-1185

* Fix expectations for commit TransientTransactionError tests

Since drivers internally retry commitTransaction, these tests should expect an additional command before the entire transaction is retried. The fail point must also trigger a second time to overcome the driver's internal retry behavior and allow withTransaction() to retry the entire transaction.

* Expect readConcern.afterClusterTime when startTransaction is retried

* Write op is still applied after a write concern error

This also consolidates repeated test content with YAML references.

* Omit closeConnection for writeConcernError fail points

* Skip write concern timeout test pending SPEC-1197

* PreparedTransactionInProgress TransientTransactionError requires 4.1.6

* Add test for retrying WriteConcernFailed non-timeout error

* Do not assert insertOne result when insert command may fail