fix: explicitly specify error labels to support mongos 4.3
Mongos does not currently fill in `errorLabels` for a `failCommand`
set on a known error code. Because of the way our logic for
determining a retryable write error has changed in the 4.4 timeline
we would no longer be able to run these tests against mongos until
the bug is fixed in the proxy. As a stopgap, we'll explicitly
specify the error label in our `failCommand`, and revert this change
once the bug is fixed.

Related to SERVER-45939