fix: ensure viable servers are used for latency window calculation
By slightly modifying these two tests we can ensure only viable
servers are used to calculate the latency window. In the first set
the RTT values are slightly modified such that only the RSSecondary
servers are used to calculate the latency window. In the second case
we introduce an RSOther server, and ensure that its RTT is not used
for the latency window calculation.

SPEC-1694