Remove limit from find tests
In the Java driver, the getMore command sends a batchLimit of 1, which results in the cursor indicating that there may be more results available. This results in an extra killCursors command being called, which causes the event expectations to fail. Removing the limit doesn't change the tests but ensures we don't need to work around this upstream issue.