DRIVERS-1816 Test redaction of replies to security sensitive commands (#1026)
* DRIVERS-1816 Test redaction of replies to security sensitive commands

* Use strings for minServerVersions

* Add 1.5 schema and validate unified APM tests

* Allow redaction of non-sensitive replies to hello

This allows redacting replies to a hello command with speculative authentication, even when the reply does not contain any sensitive fields.

* Require redacting replies to sensitive commands

This changes the test expectation for a hello command with sensitive information. Previously, the test did not require redaction of the reply, as it does not contain sensitive information (due to speculative authentication failing). However, this makes the test useless, as it can no longer test that the redaction logic applies to hello at all. Until we're able to test successful authentication, it's sensible to require redacting the reply to ensure drivers are spec compliant.

* Use correct schema version in test