Clarify that old drivers should default to OP_QUERY handshakes (#1113)
The original intention of this paragraph was to allow newer drivers published after MongoDB 3.6 to avoid implementing OP_QUERY purely for an initial handshake. Drivers that have historically supported pre-3.6 servers should continue to use OP_QUERY to ensure that meaningful error messages are raised when connecting to an old server instead of a silently dropped connection due to an unrecognized protocol.