DRIVERS-1899: Permit optional results for unacknowledged write concerns
For write commands, this was previously inferable by existing documentation for the "acknowledged" property of Write Results. Likewise for findAndModify, comments suggested that the return value could be null. This change makes the documentation more explicit.

Also adds explicit documentation that InsertOneResult and InsertManyResult are not required, which was previously inferable.