DRIVERS-2609: Deprecate "sharded-replicaset" topology type in runOnRequirements
MongoDB 3.6+ requires all shards to be replica sets, so there is no reason use "sharded-replicaset" instead of "sharded".

The "sharded-replicaset" topology type dates back to the first version of the Unified Test Format, but it's not clear it was ever truly necessary. It appears to have only been used in conjunction with 3.6+ minServerVersion requirements.