DRIVERS-2731 support named KMS providers (#1492)
* add unified test format schema 1.18 to extend `kmsProviders`

* add CSE unified tests to Makefile

* use schema 1.18 in Makefile

* add spec test: namedKMS-createDataKey

* add spec test: namedKMS-rewrapManyDataKey

* add legacy spec test: namedKMS

To test automatic encryption for named KMS

* add spec test: namedKMS-explicit

*  add `Case 6` to prose test that TLS options are applied to named providers

* remove out-dated info

AWS and local are not the only KMS providers supported. Remove sentence. Specifics of KMS providers are described in the `kmsProviders` section.

* add terms for KMS providers, KMS provider, KMS provider type, and KMS provider name

* replace usage of KMSProviderName with KMSProvider

KMSProviderName was previously one of the strings "aws", "azure", "gcp", "local", or "kmip". The term "KMS provider" consists of the same set of strings, but also includes an optional name suffix. Use KMSProvider to match the term and avoid ambiguity with "KMS provider name".

* note which KMS providers support automatic credentials

* revise terminology in `masterKey`

Note that "aws" is a `KMS provider type`, not a `KMS provider`.

* describe `KMSProviders` interface

May help readers unfamiliar with typescript interfaces.

* update CHANGELOG, Last Modified, and Version

* add test `rewrap from aws:name1 to aws:name2`

* note location of test credentials

* note requirement for libmongocrypt 1.9.0

* add rationale for no on-demand KMS credentials support for named KMS

* note unnamed and `*:name1` KMS providers use the same credentials, and `aws:name2` and `aws:name1` are different AWS accounts.