DRIVERS-2768: Restore missing "type" option docs for search indexes (#1650)
b746fcc71ad7c1b376482852c3c9cbba4eed7d26 incorrectly added the "type" option to IndexOptions instead of SearchIndexOptions. Moreover, that change was made to the RST file after it was converted to Markdown and shortly before it was replaced with a stub pointing to the Markdown file. As a result, the change was entirely lost.

This adds back documentation for the "type" option and also restructures the SearchIndexModel and SearchIndexOptions definitions to be more consistent with IndexModel and IndexOptions. The "name" option is no longer duplicated between both interfaces.