Use "SuperFences" extension to fix rendering of code blocks inside lists
Here's a test case for indented nested lists and code blocks:
https://gist.github.com/mdbmes/34e9ab8d28e53cb02556941537959469

On GitHub, indents of 2, 3, or 4 spaces all generate nested lists.
On mkdocs, the third list (4 spaces) is indented and the others are not.

The built-in fenced code block rendering in mkdocs displays the root-level
code snippets correclty, but the indented code suffers from formatting
problems: it's indented incorrectly, the background has gaps, and the
language tag is treated as part of the code.

All of these issues are addressed by the replacement fenced code
implementation in SuperFences here:
https://github.com/facelessuser/pymdown-extensions

This patch enables a pinned version of SuperFences in our mkdocs config.
It also makes a minor update to the README to direct users at the pip
requirements file.