css: undo special case for margin below nested lists
bootstrap includes a special case rule to collapse the margin below nested lists; but this doesn't help in our mkdocs-generated HTML, where lists with a mixture of paragraphs and nested lists appear to be spaced inconsistently.
This patch un-does the special case rule by setting margins for nested lists back to the default, which is the same as paragraph margins.