css: alternative fix for paragraph-in-list spacing
Removing the override helps on lists that contain mostly paragraphs, but it's not an improvement for lists that contain mostly single lines of text.
This is an alternate solution that's closer to how github's css works: paragraphs inside lists are specifically given top margin, and all list items get a quarter em of minimum top margin.